+++ app/src/main/java/com/example/nicol/elink/State/StateProyectoFinanciado.java
@@ -10,7 +10,7 @@ public class StateProyectoFinanciado extends StateProyecto {
     }
 
     @Override
-    public String financiar(int monto) {
+    public String financiar(double monto) {
         //si el proyecto esta financiado se le cobra una comision de 5% del monto ingresado al inversor
         this.proyecto.setFinanciacionActual(this.proyecto.getFinanciacionActual() * 0.95 + monto);
         return "Financiado correctamente ( 5 % de comision)";
